+++ includes/class-value-timeout-node.php
@@ -22,9 +22,7 @@
 
 namespace Newspack_Nodes;
 
-if ( ! \defined( 'ABSPATH' ) ) {
-	exit;
-}
+\defined( 'ABSPATH' ) || exit;
 
 /**
  * ValueTimeout node.